+++ src/editor.rs
@@ -1,63 +1,14 @@
 use sl_console::{self, color};
 use std::cmp;
-use std::fmt;
 use std::io;
 
 use super::complete::Completer;
 use crate::context::ColorClosure;
 use crate::event::*;
+use crate::prompt::Prompt;
 use crate::Buffer;
 use crate::History;
 use crate::{util, Term};
-
-/// User-defined prompt.
-///
-/// # Examples
-///
-/// You simply define a static prompt that holds a string.
-/// The prefix and suffix fields are intended for keybinds to change the
-/// prompt (ie the mode in vi).
-/// ```
-/// # use sl_liner::Prompt;
-/// let prompt = Prompt::from("prompt$ ");
-/// assert_eq!(&prompt.to_string(), "prompt$ ");
-/// ```
-pub struct Prompt {
-    pub prefix: Option<String>,
-    pub prompt: String,
-    pub suffix: Option<String>,
-}
-
-impl Prompt {
-    /// Constructs a static prompt.
-    pub fn from<P: Into<String>>(prompt: P) -> Self {
-        Prompt {
-            prefix: None,
-            prompt: prompt.into(),
-            suffix: None,
-        }
-    }
-
-    pub fn prefix(&self) -> &str {
-        match &self.prefix {
-            Some(prefix) => prefix,
-            None => "",
-        }
-    }
-
-    pub fn suffix(&self) -> &str {
-        match &self.suffix {
-            Some(suffix) => suffix,
-            None => "",
-        }
-    }
-}
-
-impl fmt::Display for Prompt {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}{}{}", self.prefix(), self.prompt, self.suffix())
-    }
-}
 
 /// Represents the position of the cursor relative to words in the buffer.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -1008,6 +959,7 @@ impl<'a> From<Editor<'a>> for String {
 mod tests {
     use super::*;
     use crate::context::get_buffer_words;
+    use crate::prompt::Prompt;
     use crate::History;
 
     #[test]
