+++ src/keymap/emacs.rs
@@ -192,8 +192,7 @@ fn emacs_move_word(ed: &mut Editor, direction: EmacsMoveDir) -> io::Result<()> {
 mod tests {
     use super::*;
     use crate::context::get_buffer_words;
-    use crate::editor::Prompt;
-    use crate::{Completer, Editor, History, KeyMap};
+    use crate::{Completer, Editor, History, KeyMap, Prompt};
     use sl_console::event::Key;
 
     fn simulate_key_codes<'a, 'b, M: KeyMap, I>(
